Refactor UpdateDeployment
I attempted to reduce the complexity of the #perform method by relying
on the class to encapsulate its dependencies.  Since this class is only
used to call #perform once, this is similar to extracting a method
object for the #perform method.  All unit tests pass.  I made all the
new methods private to prevent clients from calling them directly.  I
also made an attempt to order them by relevance, though I am open to
suggestions for reordering.  This class has a lot of responsibility and
in future refactorings it might be a good idea to break it up into
smaller objects with fewer responsibilities.